Separate user methods from the base client
